app/requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coverage==7.4.3
cryptography==41.0.7
dataclasses-json==0.6.4
distro==1.9.0
environs==10.0.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
frozenlist==1.4.1
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.7
langchain-community==0.0.20
langchain-core==0.1.30
langchain-openai==0.0.8
langsmith==0.1.23
loguru==0.7.2
magic-filter==1.0.12
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
oauthlib==3.2.2
openai==1.12.0
openpyxl==3.0.10
orjson==3.9.15
packaging==23.2
pandas==2.1.4
pdfminer.six==20221105
pdfplumber==0.10.3
pgvector==0.2.5
Pillow==10.1.0
pluggy==1.4.0
protobuf==4.25.2
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.1
pypdfium2==4.25.0
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.0
yarl==1.9.4
